Started commenting out and removing no longer needed code. Create stubs for validation parser.
